feat: Mobox competition
