+++ src/config/constants/trading-competition/prizes.ts
@@ -22,15 +22,191 @@ interface Config {
   [key: string]: Rank[]
 }
 
-const prizes: Config = {
+// const prizes: Config = {
+//   1: [
+//     {
+//       group: '4',
+//       rank: '1',
+//       tier: Tiers.GOLD,
+//       tokenPrizeInUsd: 12600,
+//       achievements: {
+//         image: 'fan-token-champion-gold.svg',
+//         points: 2500,
+//       },
+//       hasNft: true,
+//     },
+//     {
+//       group: '3',
+//       rank: '2 ~ 10',
+//       tier: Tiers.SILVER,
+//       tokenPrizeInUsd: 29400,
+//       achievements: {
+//         image: 'fan-token-top-10-gold.svg',
+//         points: 1250,
+//       },
+//       hasNft: true,
+//     },
+//     {
+//       group: '2',
+//       rank: '11 ~ 100',
+//       tier: Tiers.BRONZE,
+//       tokenPrizeInUsd: 25200,
+//       achievements: {
+//         image: 'fan-token-top-100-gold.svg',
+//         points: 1000,
+//       },
+//       hasNft: true,
+//     },
+
+//     {
+//       group: '1',
+//       rank: '101 ~ 500',
+//       tier: Tiers.PURPLE,
+//       tokenPrizeInUsd: 16800,
+//       achievements: {
+//         image: 'fan-token-top-500-gold.svg',
+//         points: 850,
+//       },
+//       hasNft: true,
+//     },
+
+//     {
+//       group: '0',
+//       rank: '501+',
+//       tier: Tiers.TEAL,
+//       tokenPrizeInUsd: 0,
+//       achievements: {
+//         image: 'fan-token-participant-gold.svg',
+//         points: 500,
+//       },
+//       hasNft: false,
+//     },
+//   ],
+//   2: [
+//     {
+//       group: '4',
+//       rank: '1',
+//       tier: Tiers.GOLD,
+//       tokenPrizeInUsd: 3600,
+//       achievements: {
+//         image: 'fan-token-champion-silver.svg',
+//         points: 2250,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '3',
+//       rank: '2 ~ 10',
+//       tier: Tiers.SILVER,
+//       tokenPrizeInUsd: 8400,
+//       achievements: {
+//         image: 'fan-token-top-10-silver.svg',
+//         points: 1000,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '2',
+//       rank: '11 ~ 100',
+//       tier: Tiers.BRONZE,
+//       tokenPrizeInUsd: 7200,
+//       achievements: {
+//         image: 'fan-token-top-100-silver.svg',
+//         points: 850,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '1',
+//       rank: '101 ~ 500',
+//       tier: Tiers.PURPLE,
+//       tokenPrizeInUsd: 4800,
+//       achievements: {
+//         image: 'fan-token-top-500-silver.svg',
+//         points: 500,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '0',
+//       rank: '501+',
+//       tier: Tiers.TEAL,
+//       tokenPrizeInUsd: 0,
+//       achievements: {
+//         image: 'fan-token-participant-silver.svg',
+//         points: 250,
+//       },
+//       hasNft: false,
+//     },
+//   ],
+//   3: [
+//     {
+//       group: '4',
+//       rank: '1',
+//       tier: Tiers.GOLD,
+//       tokenPrizeInUsd: 1800,
+//       achievements: {
+//         image: 'fan-token-champion-bronze.svg',
+//         points: 2100,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '3',
+//       rank: '2 ~ 10',
+//       tier: Tiers.SILVER,
+//       tokenPrizeInUsd: 4200,
+//       achievements: {
+//         image: 'fan-token-top-10-bronze.svg',
+//         points: 850,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '2',
+//       rank: '11 ~ 100',
+//       tier: Tiers.BRONZE,
+//       tokenPrizeInUsd: 3600,
+//       achievements: {
+//         image: 'fan-token-top-100-bronze.svg',
+//         points: 500,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '1',
+//       rank: '101 ~ 500',
+//       tier: Tiers.PURPLE,
+//       tokenPrizeInUsd: 2400,
+//       achievements: {
+//         image: 'fan-token-top-500-bronze.svg',
+//         points: 250,
+//       },
+//       hasNft: false,
+//     },
+//     {
+//       group: '0',
+//       rank: '501+',
+//       tier: Tiers.TEAL,
+//       tokenPrizeInUsd: 0,
+//       achievements: {
+//         image: 'fan-token-participant-bronze.svg',
+//         points: 100,
+//       },
+//       hasNft: false,
+//     },
+//   ],
+// }
+
+const mboxPrizes: Config = {
   1: [
     {
       group: '4',
       rank: '1',
       tier: Tiers.GOLD,
-      tokenPrizeInUsd: 12600,
+      tokenPrizeInUsd: 4200,
       achievements: {
-        image: 'fan-token-champion-gold.svg',
+        image: 'MBOX-champion-gold.svg',
         points: 2500,
       },
       hasNft: true,
@@ -39,9 +215,9 @@ const prizes: Config = {
       group: '3',
       rank: '2 ~ 10',
       tier: Tiers.SILVER,
-      tokenPrizeInUsd: 29400,
+      tokenPrizeInUsd: 9800,
       achievements: {
-        image: 'fan-token-top-10-gold.svg',
+        image: 'MBOX-top-500-gold.svg',
         points: 1250,
       },
       hasNft: true,
@@ -50,9 +226,9 @@ const prizes: Config = {
       group: '2',
       rank: '11 ~ 100',
       tier: Tiers.BRONZE,
-      tokenPrizeInUsd: 25200,
+      tokenPrizeInUsd: 8400,
       achievements: {
-        image: 'fan-token-top-100-gold.svg',
+        image: 'MBOX-top-500-gold.svg',
         points: 1000,
       },
       hasNft: true,
@@ -62,12 +238,12 @@ const prizes: Config = {
       group: '1',
       rank: '101 ~ 500',
       tier: Tiers.PURPLE,
-      tokenPrizeInUsd: 16800,
+      tokenPrizeInUsd: 5600,
       achievements: {
-        image: 'fan-token-top-500-gold.svg',
+        image: 'MBOX-top-500-gold.svg',
         points: 850,
       },
-      hasNft: true,
+      hasNft: false,
     },
 
     {
@@ -76,7 +252,7 @@ const prizes: Config = {
       tier: Tiers.TEAL,
       tokenPrizeInUsd: 0,
       achievements: {
-        image: 'fan-token-participant-gold.svg',
+        image: 'MBOX-participant-gold.svg',
         points: 500,
       },
       hasNft: false,
@@ -87,42 +263,42 @@ const prizes: Config = {
       group: '4',
       rank: '1',
       tier: Tiers.GOLD,
-      tokenPrizeInUsd: 3600,
+      tokenPrizeInUsd: 1200,
       achievements: {
-        image: 'fan-token-champion-silver.svg',
+        image: 'MBOX-champion-silver.svg',
         points: 2250,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '3',
       rank: '2 ~ 10',
       tier: Tiers.SILVER,
-      tokenPrizeInUsd: 8400,
+      tokenPrizeInUsd: 2800,
       achievements: {
-        image: 'fan-token-top-10-silver.svg',
+        image: 'MBOX-top-500-silver.svg',
         points: 1000,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '2',
       rank: '11 ~ 100',
       tier: Tiers.BRONZE,
-      tokenPrizeInUsd: 7200,
+      tokenPrizeInUsd: 2400,
       achievements: {
-        image: 'fan-token-top-100-silver.svg',
+        image: 'MBOX-top-500-silver.svg',
         points: 850,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '1',
       rank: '101 ~ 500',
       tier: Tiers.PURPLE,
-      tokenPrizeInUsd: 4800,
+      tokenPrizeInUsd: 1600,
       achievements: {
-        image: 'fan-token-top-500-silver.svg',
+        image: 'MBOX-top-500-silver.svg',
         points: 500,
       },
       hasNft: false,
@@ -133,7 +309,7 @@ const prizes: Config = {
       tier: Tiers.TEAL,
       tokenPrizeInUsd: 0,
       achievements: {
-        image: 'fan-token-participant-silver.svg',
+        image: 'MBOX-participant-silver.svg',
         points: 250,
       },
       hasNft: false,
@@ -144,42 +320,42 @@ const prizes: Config = {
       group: '4',
       rank: '1',
       tier: Tiers.GOLD,
-      tokenPrizeInUsd: 1800,
+      tokenPrizeInUsd: 600,
       achievements: {
-        image: 'fan-token-champion-bronze.svg',
+        image: 'MBOX-champion-bronze.svg',
         points: 2100,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '3',
       rank: '2 ~ 10',
       tier: Tiers.SILVER,
-      tokenPrizeInUsd: 4200,
+      tokenPrizeInUsd: 1400,
       achievements: {
-        image: 'fan-token-top-10-bronze.svg',
+        image: 'MBOX-top-500-bronze.svg',
         points: 850,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '2',
       rank: '11 ~ 100',
       tier: Tiers.BRONZE,
-      tokenPrizeInUsd: 3600,
+      tokenPrizeInUsd: 1200,
       achievements: {
-        image: 'fan-token-top-100-bronze.svg',
+        image: 'MBOX-top-500-bronze.svg',
         points: 500,
       },
-      hasNft: false,
+      hasNft: true,
     },
     {
       group: '1',
       rank: '101 ~ 500',
       tier: Tiers.PURPLE,
-      tokenPrizeInUsd: 2400,
+      tokenPrizeInUsd: 800,
       achievements: {
-        image: 'fan-token-top-500-bronze.svg',
+        image: 'MBOX-top-500-bronze.svg',
         points: 250,
       },
       hasNft: false,
@@ -190,7 +366,7 @@ const prizes: Config = {
       tier: Tiers.TEAL,
       tokenPrizeInUsd: 0,
       achievements: {
-        image: 'fan-token-participant-bronze.svg',
+        image: 'MBOX-participant-bronze.svg',
         points: 100,
       },
       hasNft: false,
@@ -198,4 +374,4 @@ const prizes: Config = {
   ],
 }
 
-export default prizes
+export default mboxPrizes
