+++ src/config/localization/translations.json
@@ -1602,5 +1602,9 @@
   "I acknowledge that:": "I acknowledge that:",
   "I understand that small orders are executed at higher execution price due to gas fees.": "I understand that small orders are executed at higher execution price due to gas fees.",
   "I understand that using taxed tokens on limit orders is highly discouraged and not recommended - Taxed tokens can be stuck in limit orders and lost forever because of impossibility to execute an order for such tokens": "I understand that using taxed tokens on limit orders is highly discouraged and not recommended - Taxed tokens can be stuck in limit orders and lost forever because of impossibility to execute an order for such tokens",
-  "Apply to run an IFO!": "Apply to run an IFO!"
+  "Apply to run an IFO!": "Apply to run an IFO!",
+  "Binance Fan token Trading Competition": "Binance Fan token Trading Competition",
+  "Easter Trading Competition": "Easter Trading Competition",
+  "Mobox Trading Competition": "Mobox Trading Competition",
+  "$80,000 in Prizes with Tokens and NFTs!": "$80,000 in Prizes with Tokens and NFTs!"
 }
