--- src/pages/competition.tsx
@@ -1,3 +1,0 @@
-import TradingCompetition from '../views/TradingCompetition'
-
-export default TradingCompetition
