+++ src/pages/competition/finished.tsx
@@ -1,0 +1,3 @@
+import FinishedCompetitions from '../../views/TradingCompetition/FinishedCompetitions'
+
+export default FinishedCompetitions
