+++ src/pages/competition/index.tsx
@@ -1,0 +1,3 @@
+import TradingCompetition from '../../views/TradingCompetition'
+
+export default TradingCompetition
