+++ src/views/TradingCompetition/components/BattleBanner/index.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import styled from 'styled-components'
 import { useTranslation } from 'contexts/Localization'
 import useTheme from 'hooks/useTheme'
-import AllBunniesImage from '../../pngs/all-bunnies.png'
+import AllBunniesImage from '../../pngs/mbox-all-bunnies.png'
 import { Heading1Text, Heading2Text } from '../CompetitionHeadingText'
 
 const TextStyles = (theme) => `
@@ -17,7 +17,10 @@ const ImageWrapper = styled.div`
   width: 75%;
   margin: 0 auto;
   ${({ theme }) => theme.mediaQueries.md} {
-    display: none;
+    position: absolute;
+    right: -20%;
+    bottom: -5%;
+    z-index: -1;
   }
 `
 
@@ -31,7 +34,10 @@ const StyledHeading1Text = styled(Heading1Text)`
 `
 
 const StyledHeading2Text = styled(Heading2Text)`
-  ${({ theme }) => TextStyles(theme)}
+  ${({ theme }) => TextStyles(theme)};
+  white-space: initial;
+  -webkit-text-stroke-width: 1.2px;
+  -webkit-text-stroke-color: #462091;
 `
 
 const StyledHeading = styled(Heading)`
@@ -48,21 +54,22 @@ const BattleBanner = () => {
   return (
     <Flex flexDirection="column">
       <ImageWrapper>
-        <Image src={AllBunniesImage} alt="all the bunnies" width={1208} height={659} />
+        <Image src={AllBunniesImage} alt="all the bunnies" width={523} height={395} />
       </ImageWrapper>
       <StyledText mb="16px" color="textSubtle" bold>
-        {new Date(2020, 11).toLocaleString(locale, {
+        {new Date(2022, 3).toLocaleString(locale, {
           month: 'short',
         })}{' '}
-        14-20, 2021
+        13-19, 2022
       </StyledText>
-      <StyledHeading1Text>{t('Binance Fan Token Trading Competition')}</StyledHeading1Text>
+      <StyledHeading1Text>{t('Mobox Trading Competition')}</StyledHeading1Text>
       <StyledHeading2Text background={theme.colors.gradients.gold} $fill>
-        {t('$120,000 in Prizes!')}
+        {t('$80,000 in Prizes with Tokens and NFTs!')}
       </StyledHeading2Text>
       <StyledHeading scale="md" color={theme.isDark ? 'textSubtle' : 'inputSecondary'} mt="16px">
         {t('Compete with other teams for the highest trading volume!')}
       </StyledHeading>
+      <Flex height="100px" />
     </Flex>
   )
 }
