+++ src/views/TradingCompetition/components/Rules/index.tsx
@@ -29,9 +29,7 @@ const Rules = () => {
       <StyledCardWrapper>
         <RulesCard title={t('Trade to increase your rank')}>
           <Text textAlign="center" fontSize="14px" color="textSubtle">
-            {t(
-              'Eligible pairs: SANTOS/BNB, PORTO/BNB, LAZIO/BNB, SANTOS/BUSD, PORTO/BUSD, LAZIO/BUSD, CAKE/BNB and CAKE/BUSD',
-            )}
+            {t('Eligible pairs: MBOX/BNB, MBOX/BUSD, CAKE/BNB and CAKE/BUSD')}
           </Text>
         </RulesCard>
         <RulesCard title={t('Play as a team')}>
